Removed dependency on debug (#73)
Debug uses NodeJS-specific features (e.g. text color for the console), which requires using a plugin of rollup that introduces security issues. Given it has little usage across the code base, it may be removed for the time being, and to look for a replacement solution or update the rollup plugin ourselves.